CSR-11 - fixing skills unlimited upgrades error caused by byte type

+++ Assets/ScriptableObjects/ProjectileConfigSO.cs
@@ -6,17 +6,17 @@ namespace Assets.ScriptableObjects
     public class ProjectileConfigSO : ScriptableObject
     {
         [field: SerializeField] public float DisapearingDuration { get; private set; } = 0.1f;
-        [SerializeField] private byte _startDamage;
+        [SerializeField] private int _startDamage;
         [SerializeField] private float _startSize;
         [SerializeField] private float _startSpeed;
         [SerializeField] private float _startRange;
-        [SerializeField] private byte _startMaxPiercing;
+        [SerializeField] private int _startMaxPiercing;
 
-        public byte Damage { get; set; }
+        public int Damage { get; set; }
         public float Speed { get; set; }
         public float Size { get; set; }
         public float Range { get; set; }
-        public byte MaxPiercing { get; set; }
+        public int MaxPiercing { get; set; }
 
         private void OnEnable()
         {
